+++ src/components/assessment/__tests__/index.tsx
@@ -1,62 +1,62 @@
-import { shallow } from 'enzyme'
+import { mount } from 'enzyme'
 import * as React from 'react'
+import { MemoryRouter } from 'react-router'
 
-import { mockAssessments } from '../../../mocks/assessmentAPI'
-import Assessment, { AssessmentProps } from '../index'
+import { mockAssessmentOverviews } from '../../../mocks/assessmentAPI'
+import { mockRouterProps } from '../../../mocks/components'
+import AssessmentListing, { IAssessmentListingProps } from '../AssessmentListing'
+import { AssessmentCategories } from '../assessmentShape'
 
-const defaultProps: AssessmentProps = {
-  activeTab: 0,
-  assessmentId: 0,
-  editorWidth: '50%',
-  handleAssessmentFetch: (assessmentId: number) => {},
-  handleChangeActiveTab: (activeTab: number) => {},
-  handleChapterSelect: (chapter: any, changeEvent: any) => {},
-  handleEditorEval: () => {},
-  handleEditorValueChange: (val: string) => {},
-  handleEditorWidthChange: (widthChange: number) => {},
-  handleInterruptEval: () => {},
-  handleReplEval: () => {},
-  handleReplOutputClear: () => {},
-  handleReplValueChange: (newValue: string) => {},
-  handleSideContentHeightChange: (heightChange: number) => {},
-  isRunning: false,
-  output: [],
-  questionId: 0,
-  replValue: ''
+const defaultProps: IAssessmentListingProps = {
+  assessmentCategory: AssessmentCategories.Mission,
+  assessmentOverviews: undefined,
+  handleAssessmentOverviewFetch: () => {},
+  handleResetAssessmentWorkspace: () => {},
+  handleUpdateCurrentAssessmentId: (assessmentId: number, questionId: number) => {},
+  ...mockRouterProps('/academy/missions', {})
 }
 
-const mockUndefinedAssessmentProps: AssessmentProps = {
-  ...defaultProps
-}
-
-const mockProgrammingAssessmentProps: AssessmentProps = {
+const mockUndefinedAssessmentListing: IAssessmentListingProps = {
   ...defaultProps,
-  assessment: mockAssessments[0],
-  assessmentId: 0,
-  questionId: 0
+  assessmentOverviews: undefined
 }
 
-const mockMcqAssessmentProps: AssessmentProps = {
+const mockEmptyAssessmentListing: IAssessmentListingProps = {
   ...defaultProps,
-  assessment: mockAssessments[0],
-  assessmentId: 0,
-  questionId: 2
+  assessmentOverviews: []
 }
 
-test('Assessment page "loading" content renders correctly', () => {
-  const app = <Assessment {...mockUndefinedAssessmentProps} />
-  const tree = shallow(app)
+const mockPresentAssessmentListing: IAssessmentListingProps = {
+  ...defaultProps,
+  assessmentOverviews: mockAssessmentOverviews
+}
+
+test('AssessmentListing page "loading" content renders correctly', () => {
+  const app = (
+    <MemoryRouter initialEntries={['/unknown']}>
+      <AssessmentListing {...mockUndefinedAssessmentListing} />
+    </MemoryRouter>
+  )
+  const tree = mount(app)
   expect(tree.debug()).toMatchSnapshot()
 })
 
-test('Assessment page with programming question renders correctly', () => {
-  const app = <Assessment {...mockProgrammingAssessmentProps} />
-  const tree = shallow(app)
+test('AssessmentListing page with 0 missions renders correctly', () => {
+  const app = (
+    <MemoryRouter initialEntries={['/unknown']}>
+      <AssessmentListing {...mockEmptyAssessmentListing} />
+    </MemoryRouter>
+  )
+  const tree = mount(app)
   expect(tree.debug()).toMatchSnapshot()
 })
 
-test('Assessment page with MCQ question renders correctly', () => {
-  const app = <Assessment {...mockMcqAssessmentProps} />
-  const tree = shallow(app)
+test('AssessmentListing page with multiple loaded missions renders correctly', () => {
+  const app = (
+    <MemoryRouter initialEntries={['/unknown']}>
+      <AssessmentListing {...mockPresentAssessmentListing} />
+    </MemoryRouter>
+  )
+  const tree = mount(app)
   expect(tree.debug()).toMatchSnapshot()
 })
